refactor(#792): split rig_runtime_tests into base + scene suite (sweep)
rig_runtime_tests.rs 607 -> 378, + rig_runtime_tests_scene.rs 234 (both under cap;
off debt). The #454-T5 scene-switch tests move to the sibling; the shared fixtures
(in_ep/out_ep/binding/fx/rig/input + SR const + arc_for) become pub(super) and the
new file imports them via super::tests. engine tests (36 groups) pass, zero warnings.
