+++ scripts/validate.sh
@@ -62,7 +62,6 @@ crates/block-core/src/param_tests.rs
 crates/block-mod/src/lib_tests.rs
 crates/engine/src/audio_alloc_invariant_tests.rs
 crates/engine/src/audio_signal_integrity_tests.rs
-crates/engine/src/rig_runtime_tests.rs
 crates/engine/src/runtime_tests.rs
 crates/engine/src/volume_invariants_tests.rs
 crates/engine/tests/issue_670_beat_it_real_rig.rs
